requirements.txt
Flask==2.3.3
gunicorn
requests
PyYAML==6.0.1
resend
python-dotenv